포인트 결제 기능 추가 및 Product unitCost Integer로 변경

+++ src/main/java/com/nhnacademy/shoppingmall/domain/ProductDto.java
@@ -10,5 +10,5 @@ public class ProductDto {
     private String productImage;
     private String description;
     private String modelName;
-    private Long unitCost;
+    private Integer unitCost;
 }
